refactor(installer): split mcp_config.go into 2 focused files
